pyproject.toml
[tool.poetry]
name = "slack-stockpiler"
version = "0.1.0"
description = ""
authors = ["Ryan O'Sullivan <osullivryan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
slackclient = "^2.6.2"

[tool.poetry.dev-dependencies]
pylint = "^2.5.2"
pytest = "^5.4.3"
flake8 = "^3.8.2"
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
